requirements.txt
confluent_kafka
fastavro
django
djangorestframework